+++ src/Components/common/math/discrete_points_smoothing/fem_pos_deviation_sqp_osqp_interface.cc
@@ -21,7 +21,7 @@ bool FemPosDeviationSqpOsqpInterface::Solve() {
     return false;
   }
 
-  if (ref_points_.size() > std::numeric_limits<int>::max()) {
+  if (ref_points_.size() > std::numeric_limits<size_t>::max()) {
     LOG(ERROR) << "ref_points size too large, solver early terminates";
     return false;
   }
